Fixed empty live post error
If there hadn't been any time_entries that day, @today_time_entries would be empty and throw an error in rendering the homepage. This fix added a check to the live_post method to see if it was empty. If it is empty, it returns nil and therefore doesn't have a 'live_post' and folows that logic tree.