Changed timezone and therefore Toggl API call
The website was set to use UTC but meant that when the local time changed to BST all time was out by 1 hour. In configuring the timezone in application.rb, it then changed the format of running the iso8601 method on the time. This new format was incompatible with the Toggl API.

Fixed by adding a call to the .utc method before the iso8601 method for the API call only. The returning times from the API call are correct as the time is configured in the Toggl system.

In order to show the correct time I needed to change all other calls within the app to Time.zone.now